requirements.txt
cassandra-driver
pymongo